[GSK-1473] Enabling rewrite_prompt for giskard evaluator (#1266)
* refactoring of giskard_evaluator

* Added unwrap_python_model_from_pyfunc

* introducing unwrap_python_model_from_pyfunc for langchain models

* added pyfunc model

* fixed test_infer_giskard_cls.py

* fixed format

* implemented MD's feedback

* updated

---------
