+++ python-client/tests/models/automodel/test_infer_giskard_cls.py
@@ -5,7 +5,6 @@ from giskard.models.huggingface import HuggingFaceModel
 from giskard.models.pytorch import PyTorchModel
 from giskard.models.sklearn import SKLearnModel
 from giskard.models.tensorflow import TensorFlowModel
-from giskard.models.base import CloudpickleSerializableModel
 
 
 def pytorch_model():
@@ -44,6 +43,7 @@ def pyfunc_model():
     )
 
     import mlflow
+
     model_uri = mlflow.tensorflow.log_model(tf_model, "tensorflow_model").model_uri
     return mlflow.pyfunc.load_model(model_uri)
 
@@ -72,6 +72,3 @@ def test_infer_giskard_cls(german_credit_raw_model, german_credit_catboost_raw_m
 
     giskard_cls = _infer_giskard_cls(huggingface_model())
     assert giskard_cls == HuggingFaceModel
-
-    giskard_cls = _infer_giskard_cls(pyfunc_model())
-    assert giskard_cls == CloudpickleSerializableModel
